+++ spec/redis/time_series_spec.rb
@@ -396,4 +396,29 @@ RSpec.describe Redis::TimeSeries do
       end
     end
   end
+
+  describe 'equality' do
+    let(:other_ts) { described_class.new(other_key, redis: redis) }
+
+    context 'when key and client match' do
+      let(:other_key) { ts.key }
+      let(:redis) { ts.redis }
+
+      it { is_expected.to eq other_ts }
+    end
+
+    context 'when key does not match' do
+      let(:other_key) { 'other_key' }
+      let(:redis) { ts.redis }
+
+      it { is_expected.not_to eq other_ts }
+    end
+
+    context 'when client does not match' do
+      let(:other_key) { ts.key }
+      let(:redis) { Redis.new }
+
+      it { is_expected.not_to eq other_ts }
+    end
+  end
 end