+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix aggregations for TS.RANGE command (#34)
 * Extract client handling into Client module (#32)
 * Add `uncompressed` param to TS.ADD, TS.INCRBY, TS.DECRBY (#35)
+* Add `Redis::TimeSeries::Rule` object (#38)
 
 ## 0.4.0
 * Added [hash-based filter DSL](https://github.com/dzunk/redis-time-series/tree/7173c73588da50614c02f9c89bf2ecef77766a78#filter-dsl)